Testing of shapes

+++ Shape/src/TestTeamProject/Point.java
@@ -1,10 +1,10 @@
 package TestTeamProject;
 
 public class Point {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
@@ -12,7 +12,7 @@ public class Point {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
@@ -20,7 +20,7 @@ public class Point {
         this.y = y;
     }
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
